Add palm dbt passthrough command
Many users are relying heavily on palm shell to allow control
over the dbt commands they are running.
Adding a passthrough command allows users to run whatever commands they
want inside the container.

palm dbt captures arguments as the command and allows the most common
options, with the ability to specify an option string for any additional options
that are needed